Feat: refacto SmartButton to different file
